#144 move h5_utilities to their own module
