+++ allensdk/test/core/test_brain_observatory_nwb_data_set.py
@@ -42,6 +42,8 @@ import pytest
 import os
 import h5py
 
+from test_h5_utilities import mem_h5
+
 
 NWB_FLAVORS = []
 
@@ -66,14 +68,12 @@ def data_set(request):
 
 
 @pytest.fixture
-def mem_h5(request):
-    my_file = h5py.File('my_file.h5', driver='core', backing_store=False)
-
-    def fin():
-        my_file.close()
-    request.addfinalizer(fin)
-
-    return my_file
+def stim_pres_h5(mem_h5):
+    def make_stim_pres_h5(stimulus_name):
+        mem_h5.create_group('stimulus/presentation/{}'.format(stimulus_name))
+        mem_h5.create_group('stimulus/not_presentation/{}'.format(stimulus_name))
+        return mem_h5
+    return make_stim_pres_h5
 
 
 @pytest.fixture
@@ -387,76 +387,6 @@ def test_make_repeated_indexed_time_series_stimulus_table(indexed_timeseries_h5)
     frame_dur_obt = np.array([ obt['start'].values, obt['end'].values ]).T
     assert(np.allclose( frame_dur_obt, frame_dur_exp ))
     assert(np.allclose( repeats_exp, obt['repeat'] ))
-
-
-@pytest.fixture
-def simple_h5(mem_h5):
-    mem_h5.create_group('a')
-    mem_h5.create_group('a/b')
-    mem_h5.create_group('a/b/c')
-    mem_h5.create_group('d')
-    mem_h5.create_group('a/e')
-
-    return mem_h5
-
-
-@pytest.fixture
-def simple_h5_with_datsets(simple_h5):
-    simple_h5.create_dataset(name='/a/b/c/fish', data=np.eye(10))
-    simple_h5.create_dataset(name='a/fowl', data=np.eye(15))
-    simple_h5.create_dataset(name='a/b/mammal', data=np.eye(20))
-
-    return simple_h5
-
-
-@pytest.fixture
-def stim_pres_h5(mem_h5):
-    def make_stim_pres_h5(stimulus_name):
-        mem_h5.create_group('stimulus/presentation/{}'.format(stimulus_name))
-        mem_h5.create_group('stimulus/not_presentation/{}'.format(stimulus_name))
-        return mem_h5
-    return make_stim_pres_h5
-
-
-def test_traverse_h5_file(simple_h5):
-
-    names = []
-    def cb(name, node):
-        names.append(name)
-    bonds._traverse_h5_file(cb, simple_h5)
-
-    assert( set(names) == set(['a', 'a/b', 'a/b/c', 'd', 'a/e']) )
-
-
-def test_locate_h5_objects(simple_h5):
-
-    matcher_cb = functools.partial(bonds._h5_object_matcher_relname_in, ['c', 'e'])
-    matches = bonds._locate_h5_objects(matcher_cb, simple_h5)
-
-    match_names = [ match.name for match in matches ]
-    assert( set(match_names) == set(['/a/e', '/a/b/c']) )
-
-
-def test_keyed_locate_h5_objects(simple_h5):
-
-    matcher_cbs = {
-        'e': functools.partial(bonds._h5_object_matcher_relname_in, ['e']),
-        'c': functools.partial(bonds._h5_object_matcher_relname_in, ['c']),
-    }
-
-    matches = bonds._keyed_locate_h5_objects(matcher_cbs, simple_h5)
-    assert( matches['e'].name == '/a/e' )
-    assert( matches['c'].name == '/a/b/c' )
-
-
-def test_load_datasets_by_relnames(simple_h5_with_datsets):
-
-    relnames = ['fish', 'fowl', 'mammal']
-    obt = bonds._load_datasets_by_relnames(relnames, simple_h5_with_datsets, simple_h5_with_datsets['a/b'])
-
-    assert( len(obt) == 2 )
-    assert(np.allclose( obt['fish'], np.eye(10) ))
-    assert(np.allclose( obt['mammal'], np.eye(20) ))
 
 
 def test_find_stimulus_presentation_group(stim_pres_h5):
